Add info box shortcode and enqueue stylesheet
This commit adds the main PHP file for the Info Box Plugin. The file includes the following:

- Plugin header comment block with plugin details
- Info box shortcode definition and handling
- Enqueuing the stylesheet for the info box

The shortcode `[info_box]` can be used to wrap content inside an info box. The content within the shortcode will be displayed in a boxed style format.

The stylesheet is enqueued to apply the necessary styling to the info box. The styles can be customized by modifying the `css/style.css` file.